language: setup a dictionary when loading
When setting up a language, it's useful to have a dictionary
of words available. This commit adds the necessary plumbing
to load the dictionary of the given language if there's any
present.

Right now, dictionary files are only a list of words, but in
the future we'll move to a much more sophisticated implementation
and make it a real dictionary. Each entry would have more
information such as the word type (verb, noun, etc), and any
other relevant information.

In the future we'll also implement the tools that people can
use to create and improve dictionaries themselves.