segmenter: add :language property
It is expected that more complicated segmenters would need
to access language information such as dictionaries and
word conjugators. For that, it is required that the segmenter
have access to the GwLanguage instance that created it.

This commit adds the GwSegmenter:language property, and the
necessary functions to auxiliate it.