document: introduce GwDocument
GwDocument represents a text in libgwords. It's an abstraction
layer to a complete document with one or more paragraphs, thus
GwDocument.

This commit introduces an initial prototype of the class, and
also adds all the plumbing to support retrieving the GType of
the GwDocument for a given language.

A new task helper function was added as well.